+++ make.py
@@ -82,10 +82,18 @@ if globals.arch == "x86_32":
         "-m 256 "
         "-boot menu=on,strict=on "
         "-global i8042.extended-state=off "
-        "-device piix4-ide,id=pci-ide0 "
-        "-device ide-hd,bus=pci-ide0.0,drive=hd0 "
+
+        # IDE-mode (ATA)
+        # f"-drive file={globals.DISK_PATH},if=none,id=hd0,format=raw,media=disk "
+        # "-device piix4-ide,id=pci-ide0 "
+        # "-device ide-hd,bus=pci-ide0.0,drive=hd0 "
+
+        # AHCI-mode (HBA)
+        f"-drive file={globals.DISK_PATH},if=none,id=disk0,format=raw,media=disk "
+        "-device ahci,id=ahci0 "
+        "-device ide-hd,drive=disk0,bus=ahci0.0 "
+
         "-device virtio-mouse-pci "
-        f"-drive file={globals.DISK_PATH},if=none,id=hd0,format=raw,media=disk "
         "-vga std "
         "-serial stdio "
     )